Tidy up deployer tests
